@@ -31,23 +31,24 @@
     "gulp-uglify": "^2.0.0",
     "headroom.js": "^0.9.3",
     "history": "^3.2.1",
+    "immutable": "^3.8.2",
     "keymirror": "^0.1.1",
     "lodash": "^4.17.4",
     "lodash.assign": "^4.2.0",
-    "moment": "^2.17.1",
-    "moment-timezone": "^0.5.11",
+    "moment": "^2.19.3",
+    "moment-timezone": "^0.5.14",
     "object-assign": "^4.0.1",
-    "react": "^15.4.2",
+    "react": "^15.6.2",
     "react-bootstrap-toggle": "^1.2.19",
     "react-copy-to-clipboard": "^4.2.3",
-    "react-dom": "^15.4.2",
+    "react-dom": "^15.6.2",
     "react-helmet": "^3.3.0",
     "react-player": "^0.14.1",
-    "react-router-scroll": "^0.4.1",
+    "react-router-scroll": "^0.4.4",
     "react-slick": "^0.14.5",
     "react-text-truncate": "^0.11.2",
     "react-textarea-autosize": "^4.0.5",
-    "superagent": "^3.3.1",
+    "superagent": "^3.8.2",
     "vinyl-buffer": "^1.0.0"
   },
   "devDependencies": {
@@ -62,7 +63,7 @@
     "babel-preset-stage-0": "^6.16.0",
     "babelify": "^7.3.0",
     "blue-tape": "^1.0.0",
-    "bootstrap": "^4.0.0-alpha.6",
+    "bootstrap": "^4.0.0-beta.2",
     "bootstrap-sass": "^3.3.7",
     "browser-sync": "^2.18.6",
     "browserify": "^13.3.0",
@@ -76,9 +77,9 @@
     "enzyme": "^2.7.1",
     "eslint": "^3.13.1",
     "eslint-plugin-react": "^6.9.0",
-    "express": "^4.14.0",
+    "express": "^4.16.2",
     "fastclick": "^1.0.6",
-    "flux": "^2.1.1",
+    "flux": "^3.1.3",
     "font-awesome": "^4.7.0",
     "gulp": "^3.9.0",
     "gulp-autoprefixer": "^3.1.1",
@@ -88,23 +89,24 @@
     "gulp-sass": "^3.1.0",
     "gulp-sourcemaps": "^1.10.0",
     "gulp-stylelint": "^3.9.0",
+    "Immutable": "^3.8.2",
     "jscs": "^3.0.7",
     "jsdom": "^9.12.0",
-    "mocha": "^3.2.0",
-    "node-libs-browser": "^2.0.0",
+    "mocha": "^3.5.3",
+    "node-libs-browser": "^2.1.0",
     "pre-commit": "^1.2.2",
-    "react-addons-css-transition-group": "^15.4.2",
-    "react-addons-linked-state-mixin": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
+    "react-addons-css-transition-group": "^15.6.2",
+    "react-addons-linked-state-mixin": "^15.6.2",
+    "react-addons-test-utils": "^15.6.2",
     "react-bootstrap": "^0.30.7",
     "react-hot-loader": "^1.3.1",
-    "react-intl": "^2.2.2",
-    "react-mixin": "^3.0.5",
+    "react-intl": "^2.4.0",
+    "react-mixin": "^3.1.1",
     "react-pikaday": "^0.5.1",
     "react-pure-render": "^1.0.2",
-    "react-router": "^3.0.0",
+    "react-router": "^3.2.0",
     "react-router-bootstrap": "^0.23.1",
-    "react-select": "^1.0.0-rc.2",
+    "react-select": "^1.1.0",
     "react-svg-icons": "^0.2.0",
     "react-tap-event-plugin": "^2.0.1",
     "stylelint-config-standard": "^16.0.0",
