@@ -1,7 +1,12 @@
-let Dispatcher = require("../dispatcher/Dispatcher");
-let FluxMapStore = require("flux/lib/FluxMapStore");
+import { ReduceStore } from "flux/utils";
+import Dispatcher from "../dispatcher/Dispatcher";
 
-class ElectionStore extends FluxMapStore {
+class ElectionStore extends ReduceStore {
+
+  getInitialState () {
+    return {
+    };
+  }
 
   getElectionList () {
     return this.getState().election_list || [];
